+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ## 0.5.1 (UNRELEASED)
 - [holy-lambda] Improved performance of the runtime
-- [bb layer] Renamed the Entrypoint environment variable in AWS Lambda to HL_ENTRYPOINT.
+- [bb layer] Renamed the Entrypoint environment variable in AWS Lambda to `HL_ENTRYPOINT`.
 - [bb tasks] Rename `hl:sync` to `hl:babashka-sync`. Sync downloads only `pods` and Babashka deps. Dependencies for babashka should be specified in bb.edn.
 - [bb tasks] Remove `:mvn/local-repo` from `deps.edn`. Dependencies for Clojure & Native backend are not downloaded to a local directory. 
-- [bb tasks] Java compilation happens without Docker now.
+- [bb tasks] Project compilation happens without Docker now.
 - [bb tasks] HL reverts allowance for not specifying the backend name. 
 - [bb tasks] Rename `:runtime` in `bb.edn` to `:backend` to match the documentation. 
 - [bb tasks] Remove `:entrypoint` and `:clj-alias` support. Use `:compile-cmd` in `:build` instead!
+- [docker] Holy Lambda images are now available at ghcr. Please use the new `ghcr.io/fierycod/holy-lambda-builder` instead of `fierycod/graalvm-native-image`. Images don't include `aws`, `aws sam`, and `clojure`.
 
 ## 0.5.0