+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [bb tasks] Remove `:entrypoint` and `:clj-alias` support. Use `:compile-cmd` in `:build` instead!
 - [docker] Holy Lambda images are now available at ghcr. Please use the new `ghcr.io/fierycod/holy-lambda-builder` instead of `fierycod/graalvm-native-image`. Images don't include `aws`, `aws sam`, and `clojure`.
 - [docker] Holy Lambda now supports aarch64 for all of the runtimes.
+- [holy-lambda] Fix `application/json` content type detection
 
 ## 0.5.0